requirements.txt
pandas
numpy
sqlalchemy
pymysql
streamlit
matplotlib
seaborn
pyyaml
openpyxl
jupyter
